+++ frontend/src/App.jsx
@@ -4,8 +4,8 @@ import Home from "./pages/Home";
 import Collection from "./pages/Collection";
 import Univers from "./pages/Univers";
 import Layout from "./components/Layout";
-import AddingBoardgame from "./pages/AddingBoardgame";
-import EditBoardgame from "./pages/EditBoardgame";
+import AddingBoardgame from "./pages/admin/AddingBoardgame";
+import EditBoardgame from "./pages/admin/EditBoardgame";
 import Page404 from "./pages/Page404";
 import Profil from "./pages/Profil";
 import ProtectedRoutes from "./Protections/ProtectedRoutes";
