+++ frontend/src/components/collection/CollectionCard.jsx
